infrastructure/testharness.js/add_cleanup_sync_queue.html

<!DOCTYPE HTML>
<html>
<head>
<title>Test#add_cleanup: queuing tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>

<script>
"use strict";
var firstCleaned = false;

promise_test(function(t) {
  promise_test(function() {
    assert_true(
      firstCleaned, "should not execute until first test is complete"
    );

    return Promise.resolve();
  }, "test defined when no tests are queued, but one test is executing");

  t.add_cleanup(function() {
    firstCleaned = true;
  });

  return Promise.resolve();
}, "Test with a 'cleanup' function");
</script>
</body>
</html>
